--- CarManufecturingCompany/src/srbshakib/SupplyChainManager/CustomerOrderSecneController.java
@@ -1,64 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/javafx/FXMLController.java to edit this template
- */
-package srbshakib.SupplyChainManager;
-
-import java.net.URL;
-import java.util.ResourceBundle;
-import javafx.event.ActionEvent;
-import javafx.fxml.FXML;
-import javafx.fxml.Initializable;
-
-/**
- * FXML Controller class
- *
- * @author SRB Shakib
- */
-public class CustomerOrderSecneController implements Initializable {
-
-    /**
-     * Initializes the controller class.
-     */
-    @Override
-    public void initialize(URL url, ResourceBundle rb) {
-        // TODO
-    }    
-
-    @FXML
-    private void homeButtonOnMuseClicked(ActionEvent event) {
-    }
-
-    @FXML
-    private void trackLogisticsButtonOnMuseClicked(ActionEvent event) {
-    }
-
-    @FXML
-    private void orderButtonOnMuseClicked(ActionEvent event) {
-    }
-
-    @FXML
-    private void distribruteButtonOnMuseClicked(ActionEvent event) {
-    }
-
-    @FXML
-    private void orderHistoryButtonOnMuseClicked(ActionEvent event) {
-    }
-
-    @FXML
-    private void updateInventoryButtonOnMuseClicked(ActionEvent event) {
-    }
-
-    @FXML
-    private void customerOrderButtonOnMuseClicked(ActionEvent event) {
-    }
-
-    @FXML
-    private void supplierListButtonOnMuseClicked(ActionEvent event) {
-    }
-
-    @FXML
-    private void supplyHistoryButtonOnMuseClicked(ActionEvent event) {
-    }
-    
-}
